+++ src/Graphics/Plotly/Histogram.hs
@@ -18,10 +18,10 @@ histogram nbins pts =
   let (lo, hi) = (minimum pts, maximum pts)
       binSize = (hi - lo) / realToFrac nbins
       binf :: Double -> Int
-      binf x = floor $ (x - lo) / binSize
+      binf xv = floor $ (xv - lo) / binSize
       binToX binN = realToFrac binN * binSize + lo
       bins = group $ sort $ map binf pts
-      goFill (car@(bin1,count1):cdr@((bin2,count2):_))
+      goFill (car@(bin1,_):cdr@((bin2,_):_))
          | bin2 == bin1 + 1 =  car : goFill cdr
          | otherwise = car : goFill ((bin1+1,0):cdr)
       goFill l = l