+++ changelog.md
@@ -1,5 +1,7 @@
 0.1.1
 
+* Simple module to help build traces
+
 * remove test-plotly executable